requirements-monitoring.txt
# DT-RAG v1.8.1 Monitoring System Requirements

# Core Dependencies
asyncio-mqtt>=0.13.0
asyncpg>=0.28.0
aiohttp>=3.8.5
uvicorn[standard]>=0.23.0
fastapi>=0.103.0
pydantic>=2.3.0
sqlalchemy[asyncio]>=2.0.20

# Observability and Monitoring
langfuse>=2.6.0
prometheus-client>=0.17.1
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-exporter-otlp>=1.19.0
opentelemetry-instrumentation-fastapi>=0.40b0
opentelemetry-instrumentation-sqlalchemy>=0.40b0
opentelemetry-instrumentation-aiohttp-client>=0.40b0

# System Monitoring
psutil>=5.9.5

# Redis and Caching
redis>=4.6.0
aioredis>=2.0.1

# Configuration and Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
structlog>=23.1.0

# Optional Vector Database Integrations
# qdrant-client>=1.5.0  # Uncomment if using Qdrant
# weaviate-client>=3.24.0  # Uncomment if using Weaviate
# pinecone-client>=2.2.0  # Uncomment if using Pinecone

# Development and Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation (optional)
mkdocs>=1.5.0
mkdocs-material>=9.1.0